moved safeHydrate to src root
