+++ packages/create-sitecore-jss/src/templates/nextjs/src/SafeHydrate.tsx
